+++ webchat/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "compilerOptions": {
     "target": "es5",
+    "lib": ["dom", "dom.iterable", "esnext"],
     "module": "commonjs",
     "strict": true,
     "allowJs": true,
